# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Business OS API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/client/checkouts": {
      "post": {
        "tags": [
          "client"
        ],
        "summary": "Start Checkout",
        "description": "Starts buying a plan at its current price.",
        "operationId": "start_checkout",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Checkout"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/client/checkouts/{checkout_id}": {
      "get": {
        "tags": [
          "client"
        ],
        "summary": "Get Checkout",
        "operationId": "get_checkout",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checkout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checkout Id"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Checkout"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/client/checkouts/{checkout_id}/simulate-payment": {
      "post": {
        "tags": [
          "client"
        ],
        "summary": "Simulate Payment",
        "description": "Completes a simulated checkout: records a test payment and creates the plan. Calling it\nagain returns the same plan.",
        "operationId": "simulate_payment",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checkout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checkout Id"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutPaid"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Checkout": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Plan Id"
          },
          "plan_name": {
            "type": "string",
            "title": "Plan Name"
          },
          "amount": {
            "type": "integer",
            "title": "Amount",
            "description": "Minor units"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "paid",
              "cancelled",
              "failed"
            ],
            "title": "Status"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "description": "Test payment: no money is charged"
          },
          "pay_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pay Url",
            "description": "Where to pay (real providers only)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan_id",
          "plan_name",
          "amount",
          "currency",
          "status",
          "simulated",
          "pay_url",
          "created_at"
        ],
        "title": "Checkout"
      },
      "CheckoutCreate": {
        "properties": {
          "plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Plan Id"
          }
        },
        "type": "object",
        "required": [
          "plan_id"
        ],
        "title": "CheckoutCreate"
      },
      "CheckoutPaid": {
        "properties": {
          "checkout": {
            "$ref": "#/components/schemas/Checkout"
          },
          "entitlement": {
            "$ref": "#/components/schemas/Entitlement"
          }
        },
        "type": "object",
        "required": [
          "checkout",
          "entitlement"
        ],
        "title": "CheckoutPaid"
      },
...
      "Entitlement": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          },
          "plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Plan Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "membership",
              "punch_card"
            ],
            "title": "Kind"
          },
          "credits": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits"
          },
          "credits_used": {
            "type": "integer",
            "title": "Credits Used"
          },
          "credits_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits Remaining"
          },
          "starts_on": {
            "type": "string",
            "format": "date",
            "title": "Starts On"
          },
          "ends_on": {
            "type": "string",
            "format": "date",
            "title": "Ends On"
          },
          "state": {
            "type": "string",
            "enum": [
              "active",
              "upcoming",
              "frozen",
              "used_up",
              "expired",
              "cancelled"
            ],
            "title": "State"
          },
          "price_amount": {
            "type": "integer",
            "title": "Price Amount"
          },
          "price_currency": {
            "type": "string",
            "title": "Price Currency"
          },
          "freezes": {
            "items": {
              "$ref": "#/components/schemas/Freeze"
            },
            "type": "array",
            "title": "Freezes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "receipt_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Id",
            "description": "The receipt of the payment for this plan"
          },
          "receipt_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Number"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_id",
          "plan_id",
          "name",
          "kind",
          "credits",
          "credits_used",
          "credits_remaining",
          "starts_on",
          "ends_on",
          "state",
          "price_amount",
          "price_currency",
          "freezes",
          "created_at",
          "receipt_id",
          "receipt_number"
        ],
        "title": "Entitlement"
      },
...
      "Freeze": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "starts_on": {
            "type": "string",
            "format": "date",
            "title": "Starts On"
          },
          "ends_on": {
            "type": "string",
            "format": "date",
            "title": "Ends On"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "starts_on",
          "ends_on",
          "reason"
        ],
        "title": "Freeze"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}